+++ R/glue.R
@@ -64,7 +64,7 @@ glue_data <- function(.x, ..., .sep = "", .envir = parent.frame(), .open = "{", 
   assign_args(dots[named], envir = env, data = .x)
 
   # Concatenate unnamed arguments together
-  unnamed_args <- eval_args(dots[!named], envir = env, data = .x)
+  unnamed_args <- lapply(which(!named), function(x) eval(call("force", as.symbol(paste0("..", x)))))
 
   lengths <- lengths(unnamed_args)
   if (any(lengths == 0) || length(unnamed_args) < length(dots[!named])) {